A shift in focus
Moving strictly to  21:9 and properly going through my map folder as it currently stands. There's STILL maps I missed in primary zoning.

I also want to build up a repository that not only supports the main maps but the fancy pants bonus image option as well because hell why not?

Long ass description but yeah, apologies on the sudden rugpull but I'll try and get back to previous capacity asap.