Finally merging tslashd's PR
LONG time coming. Very delayed I was originally going to go through this list, and get the images in my resolution beforehand but, spread a bit too thin with various projects right now. DEEP thanks to him and his contribution to the repo.

These will eventually be updated in 3440x1440.